kiss68030: work in progress
This works well enough to boot Linux up to the point where it tries to
write to userspace. This includes a bunch of debug code that I added to
figure out why it then starts executing rubbish. The answer is that the
MMU cannot raise a bus error when a page fault occurs.